# src/main/resources/openapi/openapi-prices-v1.yaml
openapi: 3.0.3
info:
  title: Prices API
  version: v1
servers:
  - url: http://localhost:8082
    description: Generated server url
paths:
  /api/v1/create-prices:
    post:
      tags:
        - pricesV1
      summary: Crear una promoción de precios
      description: Registra una nueva promoción para un producto en un rango de fechas determinado.
      operationId: createPrices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PricesDto'
      responses:
        '201':
          description: Promoción creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricesDto'
        '400':
          description: Solicitud inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricesDto'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricesDto'

  /api/v1/get-prices-info:
    get:
      tags:
        - pricesV1
      summary: Obtener precio aplicable
      description: Devuelve la información de precio vigente para un producto, marca y fecha proporcionados.
      operationId: pricesInformation
      parameters:
        - name: dto
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/PricesQueryDto'
      responses:
        '200':
          description: Precio encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricesDto'
        '404':
          description: No se encontró un precio válido para los parámetros proporcionados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricesDto'
        '400':
          description: Parámetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricesDto'

components:
  schemas:
    PricesDto:
      title: PricesDto
      type: object
      properties:
        brandId:
          type: integer
          format: int32
          description: Identificador único para la marca a la que pertenece el producto
          example: 1
        productId:
          type: string
          description: Identificador único para cada producto
          example: "1"
        startDate:
          type: string
          format: date-time
          description: Indica la fecha de inicio del precio y tarifa
          example: "2025-05-14T10:00:00"
        endDate:
          type: string
          format: date-time
          description: Indica la fecha de fin del precio y tarifa
          example: "2025-05-18T10:00:00"
        priceList:
          type: integer
          format: int32
          description: Identificador de una tarifa aplicada a un producto
          example: 1
        priority:
          type: integer
          format: int32
          description: Valor que indica que la tarifa con la prioridad más alta es la que aplica
          example: 1
        price:
          type: number
          format: double
          description: Indica el precio final de venta del producto
          example: 23.3
        currencyCode:
          type: string
          description: Codigo ISO de la moneda en la que se establece el precio
          example: EUR

    PricesQueryDto:
      title: PricesQueryDto
      type: object
      description: Parámetros para consultar precios vigentes
      required:
        - brandId
        - date
        - productId
      properties:
        date:
          type: string
          format: date-time
          description: Fecha en formato yyyy-MM-dd'T'HH:mm:ss.SSS
          example: "2020-06-14T18:30:00.000"
        productId:
          type: integer
          format: int32
          description: ID del producto
          example: 35455
        brandId:
          type: integer
          format: int32
          description: ID de la marca
          example: 1